+++ pages/app/meetings/index.tsx
@@ -1,4 +1,4 @@
-export { MeetingsPage as default } from '@/app/pages/app/meetings';
+export { MeetingsPage as default } from '@/pages/app/meetings';
 
 // 기존 export default를 export로 변경했습니다.
