+++ pages/app/saved-gathering.tsx
@@ -1,4 +1,4 @@
-export { SavedGatheringPage as default } from '@/app/pages/app/saved-gathering';
+export { SavedGatheringPage as default } from '@/pages/app/saved-gathering';
 
 // 기존 export default를 export로 변경했습니다.
