+++ pages/app/start.tsx
@@ -1,4 +1,4 @@
-export { StartPage as default } from '@/app/pages/app/start';
+export { StartPage as default } from '@/pages/app/start';
 
 // 기존 export default를 export로 변경했습니다.
