+++ pages/app/user/calendar.tsx
@@ -1,4 +1,4 @@
-export { CalendarPage as default } from '@/app/pages/app/user/calendar';
+export { CalendarPage as default } from '@/pages/app/user/calendar';
 
 // 기존 export default를 export로 변경했습니다.
