+++ pages/app/user/index.tsx
@@ -1,4 +1,4 @@
-export { MyPage as default } from '@/app/pages/app/user';
+export { MyPage as default } from '@/pages/app/user';
 
 // 기존 export default를 export로 변경했습니다.
