+++ pages/app/user/meeting.tsx
@@ -1,4 +1,4 @@
-export { MyMeeting as default } from '@/app/pages/app/user/meeting';
+export { MyMeeting as default } from '@/pages/app/user/meeting';
 
 // 기존 export default를 export로 변경했습니다.
