+++ pages/app/user/plan.tsx
@@ -1,4 +1,4 @@
-export { MyPlan as default } from '@/app/pages/app/user/plan';
+export { MyPlan as default } from '@/pages/app/user/plan';
 
 // 기존 export default를 export로 변경했습니다.
