+++ pages/app/user/review.tsx
@@ -1,4 +1,4 @@
-export { MyReview as default } from '@/app/pages/app/user/review';
+export { MyReview as default } from '@/pages/app/user/review';
 
 // 기존 export default를 export로 변경했습니다.
