+++ pages/auth/login/index.tsx
@@ -1,4 +1,4 @@
-export { Login as default } from '@/app/pages/auth/login';
+export { Login as default } from '@/pages/auth/login';
 
 // 기존 export default를 export로 변경했습니다.
