+++ pages/auth/login/oauth2/callback/kakao.tsx
@@ -1,4 +1,4 @@
-export { KakaoSocialLoginRedirect as default } from '@/app/pages/auth/login/oauth2/callback/kakao';
+export { KakaoSocialLoginRedirect as default } from '@/pages/auth/login/oauth2/callback/kakao';
 
 // 기존 export default를 export로 변경했습니다.
