+++ src/pages/auth/register/index.tsx
